Title: 길이 단위 환산하기, Time: 203ms, Memory: 12MB, Status: Passed - CodeTree

+++ 240831/길이 단위 환산하기/convert-length-units.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main (String args[]) {
         Scanner sc = new Scanner(System.in);
-        int ft = 30.48;
+        double ft = 30.48;
         double n = sc.nextDouble();
         System.out.printf("%.1f",n*ft);
     }
